fix: share branded cli errors
